fix collector
