--- tests/test_sphinx_builds/test_basic_run.xml
@@ -1,14 +1,0 @@
-<document source="basic_run">
-    <section classes="tex2jax_ignore mathjax_ignore" ids="a-title" names="a\ title">
-        <title>
-            a title
-        <paragraph>
-            some text
-        <container cell_index="1" cell_metadata="{}" classes="cell" exec_count="1" nb_element="cell_code">
-            <container classes="cell_input" nb_element="cell_code_source">
-                <literal_block language="ipython3" xml:space="preserve">
-                    a=1
-                    print(a)
-            <container classes="cell_output" nb_element="cell_code_output">
-                <literal_block classes="output stream" language="myst-ansi" xml:space="preserve">
-                    1
