--- tests/test_sphinx_builds/test_basic_run_md.xml
@@ -1,13 +1,0 @@
-<document source="basic_unrun">
-    <section classes="tex2jax_ignore mathjax_ignore" ids="a-title" names="a\ title">
-        <title>
-            a title
-        <paragraph>
-            this was created using 
-            <literal>
-                jupytext --to myst tests/notebooks/basic_unrun.ipynb
-        <container cell_index="1" cell_metadata="{}" classes="cell" exec_count="True" nb_element="cell_code">
-            <container classes="cell_input" nb_element="cell_code_source">
-                <literal_block language="python" xml:space="preserve">
-                    a=1
-                    print(a)
